Mudando para Single Activity e adicionando LiveData

+++ app/src/main/java/com/example/expensebook/model/entity/Entry.kt
@@ -8,7 +8,7 @@ import java.time.ZoneId
 @Entity(tableName = "entry")
 data class Entry(
     @PrimaryKey(autoGenerate = true)
-    var uid: Int? = null,
+    var uid: Long? = null,
     var date: OffsetDateTime = OffsetDateTime.now(ZoneId.systemDefault()),
     var value: Float = 0f,
     var description: String = ""
